+++ tests/testthat/test-predict.brokenstick.R
@@ -9,21 +9,21 @@ k <- length(get_knots(obj))
 
 test_that("returns proper number of rows", {
   expect_equal(nrow(predict(obj, dat)), n)
-  expect_equal(nrow(predict(obj, dat, x = NA)), m)
+  expect_equal(nrow(predict(obj, dat, x = NA, include_data = FALSE)), m)
   expect_equal(nrow(predict(obj, x = NA, y = 10)), 1L)
   expect_equal(nrow(predict(obj, x = c(NA, NA), y = c(-1, 10))), 2L)
 })
 
 test_that("returns proper number of rows with at = 'knots'", {
-  expect_equal(nrow(predict(obj, dat, x = "knots")), m * k)
-  expect_equal(nrow(predict(obj, dat, x = NA)), m)
+  expect_equal(nrow(predict(obj, dat, x = "knots", include_data = FALSE)), m * k)
+  expect_equal(nrow(predict(obj, dat, x = NA, include_data = FALSE)), m)
   expect_equal(nrow(predict(obj, x = NA, y = 10)), 1L)
 })
 
 test_that("returns proper number of rows with both data & knots", {
-  expect_equal(nrow(predict(obj, dat, x = "knots", strip_data = FALSE)), n + k * m)
-  expect_equal(nrow(predict(obj, dat, x = NA, y = 10, group = 10001, strip_data = FALSE)), 11)
-  expect_equal(nrow(predict(obj, dat, x = c(NA, NA), y = c(-1, 10), group = rep(10001, 2), strip_data = FALSE)), 12)
+  expect_equal(nrow(predict(obj, dat, x = "knots")), n + k * m)
+  expect_equal(nrow(predict(obj, dat, x = NA, y = 10, group = 10001)), 11)
+  expect_equal(nrow(predict(obj, dat, x = c(NA, NA), y = c(-1, 10), group = rep(10001, 2))), 12)
 })
 
 test_that("output = 'vector' and output = 'long' are consistent", {
@@ -44,23 +44,23 @@ test_that("output = 'vector' and output = 'long' are consistent", {
     predict(obj, dat, x = "knots", shape = "vector")
   )
   expect_equal(
-    predict(obj, dat, strip_data = FALSE)[[".pred"]],
-    predict(obj, dat, strip_data = FALSE, shape = "vector")
+    predict(obj, dat)[[".pred"]],
+    predict(obj, dat, shape = "vector")
   )
   expect_equal(
-    predict(obj, x = NA, y = 10, strip_data = FALSE)[[".pred"]],
-    predict(obj, x = NA, y = 10, strip_data = FALSE, shape = "vector")
+    predict(obj, x = NA, y = 10)[[".pred"]],
+    predict(obj, x = NA, y = 10, shape = "vector")
   )
 })
 
 exp <- fit_200
 dat <- smocc_200
 test_that("returns proper number of rows", {
-  expect_equal(nrow(predict(exp, dat, x = NA)), 200L)
-  expect_equal(nrow(predict(exp, dat, x = c(NA, NA))), 400L)
+  expect_equal(nrow(predict(exp, dat, x = NA, include_data = FALSE)), 200L)
+  expect_equal(nrow(predict(exp, dat, x = c(NA, NA), include_data = FALSE)), 400L)
   expect_equal(nrow(predict(exp, dat, x = NA, y = 1)), 1L)
   expect_equal(nrow(predict(exp, dat, x = c(NA, NA), y = c(-1, 10))), 2L)
-  expect_equal(nrow(predict(exp, dat, x = "knots")), 2200L)
+  expect_equal(nrow(predict(exp, dat, x = "knots", include_data = FALSE)), 2200L)
   expect_equal(nrow(predict(exp, dat, x = "knots", y = rep(1, 11))), 11L)
 })