+++ NEWS.md
@@ -1,3 +1,4 @@
+- Major change: Replaces the `strip_data` argument in `predict()` by the more intuitive `include_data` argument. By default, observed data are now included in the result.
 - Expression `predict(fit_light, x = "knots")` now produced the correct error message
 - Updates objects `fit_200` and `fit_200_light` to use automatic boundary (2.68y) instead of 3 yrs